requirements.txt
python-dotenv==1.0.0
jsonlines==3.1.0
playwright==1.35.0
